requirements.txt
boto3
openai
pandas
python-pptx
python-dotenv
tiktoken
pydantic
matplotlib
